Add cyclic completions bound to TAB
Now when user presses TAB, the list of completions will be cycled and
the next completion will be suggested. Currenty, the list of
completions is generated when user inserts a character or clears the
whole text entry. But not when backspace is pressed. That has to be
fixed.

The PrefixTree module was replaced with the naive Completion module.